backend/pyproject.toml
[project]
name = "dental-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.3",
    "google-genai>=1.73.0",
    "google-generativeai>=0.8.6",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.2.0",
    "python-multipart>=0.0.26",
    "ultralytics>=8.4.37",
    "uvicorn>=0.44.0",
]
